+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   retrieve data from the `id_token`. Write a new plugin configuration to take
   advantage of this feature.
 
+## [2.1.0] - 2021-05-27
+
+### Changed
+
+* Add support for mutlitenant Azure AD apps that use client credentials flow. This is done by giving an ability to override provider options to change oauth2 URLs
+
 ## [2.0.0] - 2021-04-09
 
 ### Changed
@@ -166,7 +172,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Initial release of the plugin.
 
-[Unreleased]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v1.10.1...v2.0.0
 [1.10.1]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v1.10.0...v1.10.1
 [1.10.0]: https://github.com/puppetlabs/vault-plugin-secrets-oauthapp/compare/v1.9.0...v1.10.0